add signet-crypto public crypto computation
