+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "chain-signatures/keys",
     "chain-signatures/node",
     "chain-signatures/primitives",
+    "chain-signatures/signet-crypto",
     "chain-signatures/signet-primitives",
     "integration-tests", 
     "chain-signatures/chain-integration-core", 
@@ -93,6 +94,7 @@ mpc-chain-near = { path = "./chain-signatures/chain-near" }
 mpc-keys = { path = "./chain-signatures/keys" }
 mpc-node = { path = "./chain-signatures/node" }
 mpc-primitives = { path = "./chain-signatures/primitives" }
+signet-crypto = { path = "./chain-signatures/signet-crypto" }
 signet-primitives = { path = "./chain-signatures/signet-primitives" }
 signet-program = { path = "./chain-signatures/contract-sol" }
