+++ chain-signatures/crypto/Cargo.toml
@@ -12,6 +12,7 @@ near-sdk.workspace = true
 sha3.workspace = true
 
 mpc-primitives.workspace = true
+signet-crypto.workspace = true
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 getrandom = { version = "0.2.12", features = ["custom"] }
