+++ chain-signatures/signet-crypto/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "signet-crypto"
+version.workspace = true
+edition = "2021"
+description = "Key-derivation primitives for the SigNet network"
+readme = "README.md"
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/sig-net/mpc"
+
+[dependencies]
+k256.workspace = true
+sha3.workspace = true
+signet-primitives.workspace = true
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+getrandom = { version = "0.2.12", features = ["custom"] }
